create manual module, split manual pages in manual and hints
