+++ witch_craft/src/core/mod.rs
@@ -1,5 +1,4 @@
 pub mod core;
-pub mod manual;
 pub mod messages;
 pub mod report;
 pub mod structs;
