+++ witch_craft/src/manual/mod.rs
@@ -1,0 +1,2 @@
+pub mod manual;
+pub mod hints;
